Delegate table rendering to Markout
Use Markout TableFormatter/TableMode for pretty table and TSV rendering, remove the local duplicated table formatter and Markdown table normalizer, and rely on Markout's stable TSV headers and pipe normalization.
