+++ src/dotnet-inspect/Output/OutputFormatter.cs
@@ -5,7 +5,6 @@ using DotnetInspector.Options;
 using DotnetInspector.Sections;
 using Markout;
 using Markout.Formatting;
-using System.Text;
 
 namespace DotnetInspector.Output;
 
@@ -24,8 +23,7 @@ public static class OutputFormatter
     {
         var sw = new StringWriter();
         serialize(sw, new TableFormatter(showHeader));
-        var tabular = sw.ToString();
-        return tsv ? tabular : PrettyTableFormatter.Format(tabular);
+        return sw.ToString();
     }
 
     public static void WriteTable(bool tsv, TextWriter output, bool showHeader, Action<TextWriter, IMarkoutFormatter> serialize) =>
@@ -34,55 +32,12 @@ public static class OutputFormatter
     public static MarkoutWriterOptions ConfigureTableWriterOptions(MarkoutWriterOptions options, bool tsv)
     {
         if (tsv)
-            options.FormatTableHeader = header => ToMachineHeader(header.Name);
+            options.TableMode = MarkoutTableMode.Tsv;
         return options;
     }
 
     public static MarkoutWriterOptions CreateTableWriterOptions(bool tsv) =>
         ConfigureTableWriterOptions(new MarkoutWriterOptions(), tsv);
-
-    private static string ToMachineHeader(string name)
-    {
-        if (string.IsNullOrWhiteSpace(name))
-            return "";
-
-        var sb = new StringBuilder(name.Length + 4);
-        var lastWasSeparator = true;
-
-        for (var i = 0; i < name.Length; i++)
-        {
-            var c = name[i];
-            if (!char.IsLetterOrDigit(c))
-            {
-                AppendSeparator(sb, ref lastWasSeparator);
-                continue;
-            }
-
-            if (char.IsUpper(c))
-            {
-                var previousIsLowerOrDigit = i > 0 && (char.IsLower(name[i - 1]) || char.IsDigit(name[i - 1]));
-                var nextIsLower = i + 1 < name.Length && char.IsLower(name[i + 1]);
-                if (sb.Length > 0 && !lastWasSeparator && (previousIsLowerOrDigit || nextIsLower))
-                    AppendSeparator(sb, ref lastWasSeparator);
-            }
-
-            sb.Append(char.ToLowerInvariant(c));
-            lastWasSeparator = false;
-        }
-
-        if (sb.Length > 0 && sb[^1] == '_')
-            sb.Length--;
-        return sb.ToString();
-    }
-
-    private static void AppendSeparator(StringBuilder sb, ref bool lastWasSeparator)
-    {
-        if (!lastWasSeparator && sb.Length > 0)
-        {
-            sb.Append('_');
-            lastWasSeparator = true;
-        }
-    }
 
     public static string FormatResult(InspectionResult result, InspectionOptions options,
         SectionPipeline<InspectionResult> pipeline)
